rename grpc-auth to grpcauth
