+++ internal/grpcauth/error.go
@@ -1,4 +1,4 @@
-package grpc_auth
+package grpcauth
 
 import "errors"
